recommend/requirements3.txt
emoji
Flask
numpy
annoy==1.15.2
# annoy
# git+https://github.com/spotify/annoy.git
scikit-learn
emoji
pyknp
pymongo
google-cloud-storage
pytorch-pretrained-bert
torch==1.8.0+cpu
torchvision==0.9.0+cpu
torchaudio==0.8.0
-f https://download.pytorch.org/whl/torch_stable.html

# annoy==1.17.0
# asn1crypto==0.24.0
# attrs==17.4.0
# Automat==0.6.0
# blinker==1.4
# boto3==1.17.29
# botocore==1.20.29
# certifi==2018.1.18
# chardet==3.0.4
# click==6.7
# cloud-init==20.4
# colorama==0.3.7
# command-not-found==0.3
# configobj==5.0.6
# constantly==15.1.0
# crcmod==1.7
# cryptography==2.1.4
# dataclasses==0.8
# distro-info===0.18ubuntu0.18.04.1
# ec2-hibinit-agent==1.0.0
# emoji==1.2.0
# Flask==1.1.2
# hibagent==1.0.1
# httplib2==0.9.2
# hyperlink==17.3.1
# idna==2.6
# incremental==16.10.1
# itsdangerous==1.1.0
# Jinja2==2.11.3
# jmespath==0.10.0
# joblib==1.0.1
# jsonpatch==1.16
# jsonpointer==1.10
# jsonschema==2.6.0
# keyring==10.6.0
# keyrings.alt==3.0
# language-selector==0.1
# MarkupSafe==1.0
# mecab-python3==0.996.5
# netifaces==0.10.4
# numpy==1.19.5
# oauthlib==2.0.6
# PAM==0.4.2
# Pillow==8.1.2
# pyasn1==0.4.2
# pyasn1-modules==0.2.1
# pycrypto==2.6.1
# pygobject==3.26.1
# PyJWT==1.5.3
# pyknp==0.4.6
# pymongo==3.11.3
# pyOpenSSL==17.5.0
# pyserial==3.4
# python-apt==1.6.5+ubuntu0.5
# python-dateutil==2.8.1
# python-debian==0.1.32
# pytorch-pretrained-bert==0.6.2
# pyxdg==0.25
# PyYAML==3.12
# regex==2020.11.13
# requests==2.18.4
# requests-unixsocket==0.1.5
# s3transfer==0.3.4
# scikit-learn==0.24.1
# scipy==1.5.4
# SecretStorage==2.3.1
# service-identity==16.0.0
# six==1.11.0
# ssh-import-id==5.7
# systemd-python==234
# threadpoolctl==2.1.0
# torch==1.8.0+cpu
# torchaudio==0.8.0
# torchvision==0.9.0+cpu
# tqdm==4.59.0
# Twisted==17.9.0
# typing-extensions==3.7.4.3
# ufw==0.36
# unattended-upgrades==0.1
# urllib3==1.26.4
# Werkzeug==1.0.1
# zope.interface==4.3.2
